update functionality not finished

+++ src/components/Beverages.vue
@@ -5,6 +5,7 @@
       <v-client-table :columns="columns" :data="beverages" :options="options">
         <a slot="addAmount" slot-scope="props" class="fa fa-plus fa-2x" @click="addAmount(props.row._id)"></a>
         <a slot="remove" slot-scope="props" class="fa fa-trash-o fa-2x" @click="deleteRecord(props.row._id)"></a>
+        <a slot="edit" slot-scope="props" class="fa fa-edit fa-2x" @click="editBeverage(props.row._id)"></a>
       </v-client-table>
     </div>
   </div>
@@ -25,7 +26,7 @@ export default {
       messagetitle: ' Beverages List ',
       beverages: [],
       errors: [],
-      columns: ['type', 'name', 'brand', 'size', 'amount', 'price', 'addAmount', 'remove'],
+      columns: ['type', 'name', 'brand', 'size', 'amount', 'price', 'addAmount', 'edit', 'remove'],
       options: {
         headings: {
           _id: 'ID',
@@ -44,6 +45,10 @@ export default {
     this.loadBeverages()
   },
   methods: {
+    editBeverage: function (id) {
+      this.$router.params = id
+      this.$router.push('edit')
+    },
     loadBeverages: function () {
       BeverageService.fetchBeverages()
         .then(response => {
@@ -68,14 +73,36 @@ export default {
         })
     },
     deleteRecord: function (id) {
-      BeverageService.deleteRecord(id)
-        .then(response => {
-          this.loadBeverages()
-        })
-        .catch(error => {
-          this.errors.push(error)
-          console.log(error)
-        })
+      this.$swal({
+        title: 'Are you totaly sure?',
+        text: 'You can\'t Undo this action',
+        type: 'warning',
+        showCancelButton: true,
+        confirmButtonText: 'OK Delete it',
+        cancelButtonText: 'Cancel',
+        showCloseButton: true,
+        showLoaderOnConfirm: true
+      }).then((result) => {
+        console.log('SWAL Result : ' + result)
+        if (result === true) {
+          BeverageService.deleteRecord(id)
+            .then(response => {
+              // JSON responses are automatically parsed.
+              this.message = response.data
+              console.log(this.message)
+              this.loadBeverages()
+              // Vue.nextTick(() => this.$refs.vuetable.refresh())
+              this.$swal('Deleted', 'You successfully deleted this Beverage ', 'success')
+            })
+            .catch(error => {
+              this.$swal('ERROR', 'Something went wrong trying to Delete ' + error, 'error')
+              this.errors.push(error)
+              console.log(error)
+            })
+        } else {
+          this.$swal('Cancelled', 'The beverage is still there!', 'info')
+        }
+      })
     }
   }
 }
